refactor: focusable element 관련 유틸 및 타입 분리 (#ATR-606)
