+++ packages/design-system/packages/core/src/core/types/focusableElement.ts
@@ -1,0 +1,7 @@
+export type FocusableElement =
+  | HTMLAnchorElement
+  | HTMLButtonElement
+  | HTMLInputElement
+  | HTMLTextAreaElement
+  | HTMLSelectElement
+  | HTMLDetailsElement
