+++ packages/design-system/packages/core/src/core/types/index.ts
@@ -1,1 +1,2 @@
 export * from './headingTag'
+export * from './focusableElement'
