+++ packages/design-system/packages/core/src/core/utils/index.ts
@@ -1,2 +1,3 @@
 // export { default as forwardRefWithGeneric } from './forwardRefWithGeneric'
 export { default as rem } from './rem'
+export * from './getFocusableElement'
